+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Refactored `internal/stack/record.go` and extract functions `record.extractSubstringAfterLastSlash`, `record.extractPackagePathAndName`
 * Refactored `metrics/driver.go` and extract functions
 * Refactored `internal/backoff/backoff_test.go` and extract test functions
 * Refactored `metrics/sql.go` and extract functions